Add primitives unit tests

+++ Source/NbtLib/NbtDoubleTag.cs
@@ -2,7 +2,7 @@
 {
     public class NbtDoubleTag : NbtTag
     {
-        public override NbtTagType TagType => NbtTagType.Long;
+        public override NbtTagType TagType => NbtTagType.Double;
         public double Payload { get; set; }
     }
 }
